Implemented Close Button
Currently, the date-picker only has 'Today' and 'Clear' buttons
implemented. There is no button to close (unlike the JQuery version of
the plugin, where the close button in present). This prove to be a
problem, specially in mobile devices, where the only way to close the
date-picker was to select a date or click in any of the other buttons.
With the 'Close' button, you are now able to close the date-picker with
no changes to previous selections.
Changes in the logic implemented on the corresponding JS files. Grunt
vervose runs with no errors.
Also added the corresponding classes to the less files so that the
'Close' button has it's style info.
I'm including the resulting built files as well.